refactor: 코드 컨벤션 적용
- HtmlRenderer class 메소드 분리 및 코드 컨벤션 적용
- HtmlStyleRenderer class 메소드 분리 및 코드 컨벤션 적용
- HtmlStyleRenderer 패키지 이동
